refactor(ai): split ExtractorService into llm/local/refine/runner (max-params 2/3)
ExtractorService injected 6 deps (config, registry, localPredicates,
extractionCache, localNer, extractionPatterns) + metrics. The pipeline's
collaborators are a web (cache/registry/local/llm all reused across the
skip + assembly paths), so a strict ≤3 needs a nested split:
- ExtractorLlmService (config, metrics) — OpenAI client, system-prompt
  assembly, scPasses, callLlm.
- ExtractorLocalService (localNer, extractionPatterns) — the skip-LLM
  local fast path + post-assembly pattern emission (skip flag via env).
- ExtractorRefineService (registry, localPredicates) — predicate
  override + EDC canonicalize (threshold via env).
- ExtractorRunnerService (llm, local, refine) — the engine: skip → LLM
  (single / N-pass SC) → grounding/validation → refine → persist.
- ExtractorService (extractionCache, registry, runner) — input clamp,
  snapshot load, cache memoisation; delegates to the runner.
Each class ≤3 deps. The sc-passes + prompt-injection specs build the
chain and patch the sub-services. Public extract()/modelId() unchanged.

Part of the max-params=3 god-class split program.
